test comparisons between scalars and 0-dim PDLs with badflag == 1
This is to elucidate the issues with stat() and per-PDL badvalues
brought up by Marek Gierliński in SF#390
<http://sourceforge.net/p/pdl/bugs/390/>,
<https://github.com/PDLPorters/pdl/issues/124>.

When testing the scalar comparison, test across several values on either
side of bad value so that it is clear which case is failing.

The test must be skipped if BADVAL_USENAN: the test sets `badvalue()`.