throw a warning when using comparison operator and badval is in {0, 1}
When using a comparison operator, inconsistent results can appear when a
badvalue is either 0 or 1. This is because {0, 1} is the range of all
operators that return a logical PDL. This means that if the input PDL
contains badvalues, it can not both represent badvalues that are the
result of computing the operator *and* logical values at the same time.

This is a continuation of the work addressing the problem brought up in
<http://sourceforge.net/p/pdl/bugs/390/>, <https://github.com/PDLPorters/pdl/issues/124>.