Fix to Solaris build problem re quiet_nan (use long in signature)
The quiet_nan() function on Solaris has the signature:

    double quiet_nan(long n);

and this adds a define to set the `long` parameter in the function
prototype and function definition.

It appears that the logic is already in the code,
just that the NANARG determined in mconf.h was not
used to declare the local quiet_nan version.