Fix sf#396, Module docs can not be found more than once.
If a user did 'help PDL::IO::Storable' or some other 2-deep module name,
the search for the proper module file from which to display docs
would change the reported location of the docs database (and the documentation head)
in a initialized once-per-pdl-shell variable.  Subsequent searches would
be looking for PDL/IO/IO/Storable.pm, PDL/IO/IO/IO/Storable.pm, etc,
which couldn't be found.  This was due to how foreach aliases the
loop variable to the input array.  Making an explicit copy removes
the aliasing and keeps the variable unchanged.