Change handling of empty piddles by PDL::info and 'help vars'.
When using the $piddle->info command or 'help vars' or '? vars'
from the perldl shell, both null piddles and empty piddles were
printed as 'PDL->null'. But empties and nulls are not the same:
empties have a type, and a dim list (at least 1 dim is size 0).

This change allows empties to show up in the variable listing as
such, and I added some tests.