Fixes badval propagation to support all badval configurations
The changes in <https://github.com/PDLPorters/pdl/pull/188> broke the
build for all cases of `BADVAL_USENAN` xor `BADVAL_PER_PDL`.

This was because the checking for badvalues when copying in
`pdl_kludge_copy_*` did not take into account these variations in
accessing and setting the badvalue.

For example, a simple comparison using `==` is not valid when NaN
values are being compared. Here we need to use `!finite(...)` to make
the comparison if the type can use NaN values.