Update original indadd signature
It was originally a scalar operation and
used implicit threading for the looping.
Because we need the loop index for location,
and because the original usage has always
been applied to a vector of values, it has
changed to a vector index for a and ind.
Tests still pass and this makes the way
to refactor into the coming PDL::IndOps.