Fix comparison operators in PDL::Complex
It's not that they are mathematical nonsense!  That we can deal with.
It's the fact that they were causing segmentation faults because each
comparison operator used itself, resulting in a too-deep recursion.

Now we explicitly use PDL::lt and friends to avoid the recursion.
And we use a final ternary operator to make sure the result is
not of type PDL::Complex.